+++ NEWS.md
@@ -7,7 +7,7 @@
 - Update mapbox dependency from `0.38` to `1.11`
 - Fix country name conversion in `e_country_names` [#189](https://github.com/JohnCoene/echarts4r/issues/189)
 - Fix legend toggle in radar charts [#74](https://github.com/JohnCoene/echarts4r/issues/74)
-- Expose `span` argument of `stats::loess` in `e_loess` [#194](https://github.com/JohnCoene/echarts4r/issues/194)
+- `e_loess`, and `e_lm` take a `model_args` argument to pass arguments to underlying model function [#194](https://github.com/JohnCoene/echarts4r/issues/194)
 
 # echarts4r 0.3.2